Tarea2/arquitectura-2/api/requirements.txt
litestar[standard]>=2.4.0,<3.0.0
pybase62
redis>=5.0.0
uvicorn